+++ IGI_Natives/Features.hpp
@@ -3,15 +3,5 @@
 
 void ScriptCompile();
 void StatusMsgShow();
-void StatusMsgShowText();
-void StatusMsgShowMonitorText();
-void HumanTaskViewReset();
-void PlayerXPHitShow();
-void TestWarningShow();
-void TestErrorShow();
-void TestAmmoTypeOpen();
-void TestSoundLoad();
-void TestLoadingScreenShow();
-void TestAiNatives();
 void DllMainLoop();
 void QuitLevelMain();